packages/graph-client/pyproject.toml
[project]
name = "cstack-graph-client"
version = "0.0.1"
description = "Typed Microsoft Graph client wrapper for cstack tools"
requires-python = ">=3.12"
dependencies = [
    "msgraph-sdk>=1.10",
    "azure-identity>=1.19",
    "cryptography>=43",
    "cstack-schemas",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cstack_graph_client"]
